Refactor: Move getServerAndReactClient to test helpers
