+++ packages/react-query/test/queryOptions.test.tsx
@@ -1,4 +1,4 @@
-import { getServerAndReactClient } from './__reactHelpers';
+import { getServerAndReactClient } from './__testHelpers';
 import {
   infiniteQueryOptions,
   skipToken,
