+++ packages/react-query/test/rsc-prefetch.test.tsx
@@ -1,4 +1,4 @@
-import { getServerAndReactClient } from './__reactHelpers';
+import { getServerAndReactClient } from './__testHelpers';
 import {
   defaultShouldDehydrateQuery,
   QueryClient,
